189 - template query test
