+++ dookug-testsuite/dookug-ts-common/dookug-ts-common-rest/src/main/java/hu/icellmobilsoft/dookug/ts/common/rest/mprestclient/test/IDocumentStoredTemplateTestRestClient.java
@@ -1,0 +1,38 @@
+/*-
+ * #%L
+ * DookuG
+ * %%
+ * Copyright (C) 2023 - 2025 i-Cell Mobilsoft Zrt.
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package hu.icellmobilsoft.dookug.ts.common.rest.mprestclient.test;
+
+import jakarta.ws.rs.Path;
+
+import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
+
+import hu.icellmobilsoft.dookug.api.rest.test.IDocumentStoredTemplateTest;
+import hu.icellmobilsoft.dookug.api.url.TemplatePath;
+
+/**
+ * MP REST Client interface extension on REST test interface
+ *
+ * @author mate.biro
+ * @since 2.2.0
+ */
+@RegisterRestClient
+@Path(TemplatePath.TEST_DOCUMENT_STOREDTEMPLATE)
+public interface IDocumentStoredTemplateTestRestClient extends IDocumentStoredTemplateTest {
+}
